+++ Sources/SmileID/Classes/DocumentVerification/DocumentVerificationResultDelegate.swift
@@ -1,7 +1,7 @@
 import Foundation
 
-/// The result of a document capture session
-public protocol DocumentCaptureResultDelegate: AnyObject {
+/// The result of a Document Verification
+public protocol DocumentVerificationResultDelegate: AnyObject {
     /// Delegate method called after a successful Document Verification capture and submission.
     /// It indicates that the capture and network requests were successful. The job may or may not
     /// be complete. Use `jobStatusResponse.jobComplete` to check if a job is complete and